Move Person relationships code into concern
